extract func parser helper
