+++ src/frontend/parser.ts
@@ -7,7 +7,6 @@ import type {
   BinDeclNode,
   ConstDeclNode,
   EnumDeclNode,
-  FuncDeclNode,
   HexDeclNode,
   ImmExprNode,
   ModuleFileNode,
@@ -59,6 +58,7 @@ import {
 } from './parseModuleCommon.js';
 import { parseTopLevelExternDecl } from './parseExternBlock.js';
 import { parseEnumDecl } from './parseEnum.js';
+import { parseTopLevelFuncDecl } from './parseFunc.js';
 import { parseGlobalsBlock } from './parseGlobals.js';
 import { parseOpParamsFromText, parseParamsFromText } from './parseParams.js';
 import { parseTypeDecl, parseUnionDecl } from './parseTypes.js';
@@ -308,366 +308,25 @@ export function parseModuleFile(
 
     const funcTail = consumeTopKeyword(rest, 'func');
     if (funcTail !== undefined) {
-      const exported = hasExportPrefix;
-      const header = funcTail;
-      const openParen = header.indexOf('(');
-      const closeParen = header.lastIndexOf(')');
-      if (openParen < 0 || closeParen < openParen) {
-        diagInvalidHeaderLine(
+      const parsedFunc = parseTopLevelFuncDecl(
+        funcTail,
+        text,
+        span(file, lineStartOffset, lineEndOffset),
+        lineNo,
+        i,
+        hasExportPrefix,
+        {
+          file,
+          lineCount,
           diagnostics,
           modulePath,
-          'func header',
-          text,
-          '<name>(...): <retType>',
-          lineNo,
-        );
-        i++;
-        continue;
-      }
-
-      const name = header.slice(0, openParen).trim();
-      if (!/^[A-Za-z_][A-Za-z0-9_]*$/.test(name)) {
-        diag(
-          diagnostics,
-          modulePath,
-          `Invalid func name ${formatIdentifierToken(name)}: expected <identifier>.`,
-          { line: lineNo, column: 1 },
-        );
-        i++;
-        continue;
-      }
-      if (isReservedTopLevelName(name)) {
-        diag(
-          diagnostics,
-          modulePath,
-          `Invalid func name "${name}": collides with a top-level keyword.`,
-          { line: lineNo, column: 1 },
-        );
-        i++;
-        continue;
-      }
-
-      const funcStartOffset = lineStartOffset;
-      const headerSpan = span(file, lineStartOffset, lineEndOffset);
-      const afterClose = header.slice(closeParen + 1).trimStart();
-      let returnRegs: string[] | undefined;
-      if (afterClose.length === 0) {
-        returnRegs = [];
-      } else {
-        const retMatch = /^:\s*(.+)$/.exec(afterClose);
-        if (!retMatch) {
-          diag(diagnostics, modulePath, `Invalid func header: expected ": <return registers>"`, {
-            line: lineNo,
-            column: 1,
-          });
-          i++;
-          continue;
-        }
-        const parsedRegs = parseReturnRegsFromText(
-          retMatch[1]!.trim(),
-          headerSpan,
-          lineNo,
-          diagnostics,
-          modulePath,
-        );
-        if (!parsedRegs) {
-          i++;
-          continue;
-        }
-        returnRegs = parsedRegs.regs;
-      }
-
-      const paramsText = header.slice(openParen + 1, closeParen);
-      const params = parseParamsFromText(modulePath, paramsText, headerSpan, diagnostics, {
-        isReservedTopLevelName,
-      });
-      if (!params) {
-        i++;
-        continue;
-      }
-
-      i++;
-
-      // Optional function-local `var` block; function instruction body is parsed
-      // as an implicit instruction stream.
-      let locals: VarBlockNode | undefined;
-      let asmStartOffset: number | undefined;
-      let interruptedBeforeBodyKeyword: string | undefined;
-      let interruptedBeforeBodyLine: number | undefined;
-      while (i < lineCount) {
-        const { raw: raw2, startOffset: so2 } = getRawLine(i);
-        const t2 = stripComment(raw2).trim();
-        const t2Lower = t2.toLowerCase();
-        if (t2.length === 0) {
-          i++;
-          continue;
-        }
-        const t2TopKeyword = topLevelStartKeyword(t2);
-        if (t2TopKeyword !== undefined && t2Lower !== 'var') {
-          interruptedBeforeBodyKeyword = t2TopKeyword;
-          interruptedBeforeBodyLine = i + 1;
-          break;
-        }
-
-        if (t2Lower === 'var') {
-          const varStart = so2;
-          i++;
-          const decls: VarDeclNode[] = [];
-          const declNamesLower = new Set<string>();
-          let varTerminated = false;
-
-          while (i < lineCount) {
-            const { raw: rawDecl, startOffset: soDecl, endOffset: eoDecl } = getRawLine(i);
-            const tDecl = stripComment(rawDecl).trim();
-            const tDeclLower = tDecl.toLowerCase();
-            if (tDecl.length === 0) {
-              i++;
-              continue;
-            }
-            if (tDeclLower === 'end') {
-              locals = {
-                kind: 'VarBlock',
-                span: span(file, varStart, eoDecl),
-                scope: 'function',
-                decls,
-              };
-              i++; // consume var-terminating end
-              varTerminated = true;
-              break;
-            }
-            if (tDeclLower === 'asm') {
-              diag(
-                diagnostics,
-                modulePath,
-                `Function-local var block must end with "end" before function body`,
-                { line: i + 1, column: 1 },
-              );
-              locals = {
-                kind: 'VarBlock',
-                span: span(file, varStart, soDecl),
-                scope: 'function',
-                decls,
-              };
-              i++; // consume asm so body parsing can continue
-              varTerminated = true;
-              break;
-            }
-            const tDeclTopKeyword = topLevelStartKeyword(tDecl);
-            if (tDeclTopKeyword !== undefined) {
-              if (looksLikeKeywordBodyDeclLine(tDecl)) {
-                diagInvalidBlockLine(
-                  diagnostics,
-                  modulePath,
-                  'var declaration',
-                  tDecl,
-                  '<name>: <type>',
-                  i + 1,
-                );
-                i++;
-                continue;
-              }
-              interruptedBeforeBodyKeyword = tDeclTopKeyword;
-              interruptedBeforeBodyLine = i + 1;
-              locals = {
-                kind: 'VarBlock',
-                span: span(file, varStart, soDecl),
-                scope: 'function',
-                decls,
-              };
-              break;
-            }
-
-            const declSpan = span(file, soDecl, eoDecl);
-            const parsed = parseVarDeclLine(tDecl, declSpan, i + 1, 'var', {
-              diagnostics,
-              modulePath,
-              isReservedTopLevelName,
-            });
-            if (!parsed) {
-              i++;
-              continue;
-            }
-            const localNameLower = parsed.name.toLowerCase();
-            if (declNamesLower.has(localNameLower)) {
-              diag(diagnostics, modulePath, `Duplicate var declaration name "${parsed.name}".`, {
-                line: i + 1,
-                column: 1,
-              });
-              i++;
-              continue;
-            }
-            declNamesLower.add(localNameLower);
-            decls.push(parsed);
-            i++;
-          }
-          if (interruptedBeforeBodyKeyword !== undefined) break;
-          if (!varTerminated) {
-            diag(
-              diagnostics,
-              modulePath,
-              `Unterminated func "${name}": expected "end" to terminate var block`,
-              { line: lineNo, column: 1 },
-            );
-            break;
-          }
-          continue;
-        }
-
-        if (t2Lower === 'end') {
-          asmStartOffset = so2;
-          break;
-        }
-        asmStartOffset = so2;
-        break;
-      }
-
-      if (asmStartOffset === undefined) {
-        if (interruptedBeforeBodyKeyword !== undefined && interruptedBeforeBodyLine !== undefined) {
-          diag(
-            diagnostics,
-            modulePath,
-            `Unterminated func "${name}": expected function body before "${interruptedBeforeBodyKeyword}"`,
-            { line: interruptedBeforeBodyLine, column: 1 },
-          );
-          continue;
-        }
-        diag(diagnostics, modulePath, `Unterminated func "${name}": expected function body`, {
-          line: lineNo,
-          column: 1,
-        });
-        break;
-      }
-
-      const asmItems: AsmItemNode[] = [];
-      const asmControlStack: AsmControlFrame[] = [];
-      let terminated = false;
-      let interruptedByKeyword: string | undefined;
-      let interruptedByLine: number | undefined;
-      while (i < lineCount) {
-        const { raw: rawLine, startOffset: lineOffset, endOffset } = getRawLine(i);
-        const withoutComment = stripComment(rawLine);
-        const content = withoutComment.trim();
-        const contentLower = content.toLowerCase();
-        if (content.length === 0) {
-          i++;
-          continue;
-        }
-        if (contentLower === 'asm' && asmControlStack.length === 0 && asmItems.length === 0) {
-          diag(
-            diagnostics,
-            modulePath,
-            `Unexpected "asm" in function body (function bodies are implicit)`,
-            { line: i + 1, column: 1 },
-          );
-          i++;
-          continue;
-        }
-
-        if (contentLower === 'end' && asmControlStack.length === 0) {
-          terminated = true;
-          const funcEndOffset = endOffset;
-          const funcSpan = span(file, funcStartOffset, funcEndOffset);
-          const asmSpan = span(file, asmStartOffset, funcEndOffset);
-          const asm: AsmBlockNode = { kind: 'AsmBlock', span: asmSpan, items: asmItems };
-
-          const funcNode: FuncDeclNode = {
-            kind: 'FuncDecl',
-            span: funcSpan,
-            name,
-            exported,
-            params,
-            ...(returnRegs ? { returnRegs } : {}),
-            ...(locals ? { locals } : {}),
-            asm,
-          };
-          items.push(funcNode);
-          i++;
-          break;
-        }
-        const topKeyword = topLevelStartKeyword(content);
-        if (topKeyword !== undefined) {
-          interruptedByKeyword = topKeyword;
-          interruptedByLine = i + 1;
-          break;
-        }
-
-        const fullSpan = span(file, lineOffset, endOffset);
-        const contentStart = withoutComment.indexOf(content);
-        const contentSpan =
-          contentStart >= 0
-            ? span(file, lineOffset + contentStart, lineOffset + withoutComment.length)
-            : fullSpan;
-
-        /* label: */
-        const labelMatch = /^([A-Za-z_][A-Za-z0-9_]*)\s*:\s*(.*)$/.exec(content);
-        if (labelMatch) {
-          const label = labelMatch[1]!;
-          const remainder = labelMatch[2] ?? '';
-          const labelNode: AsmLabelNode = { kind: 'AsmLabel', span: fullSpan, name: label };
-          asmItems.push(labelNode);
-          if (remainder.trim().length > 0) {
-            const stmtNode = parseAsmStatement(
-              modulePath,
-              remainder,
-              contentSpan,
-              diagnostics,
-              asmControlStack,
-            );
-            appendParsedAsmStatement(asmItems, stmtNode);
-          }
-          i++;
-          continue;
-        }
-
-        const stmtNode = parseAsmStatement(
-          modulePath,
-          content,
-          contentSpan,
-          diagnostics,
-          asmControlStack,
-        );
-        appendParsedAsmStatement(asmItems, stmtNode);
-        i++;
-      }
-
-      if (!terminated) {
-        if (interruptedByKeyword !== undefined && interruptedByLine !== undefined) {
-          for (const frame of asmControlStack) {
-            if (isRecoverOnlyControlFrame(frame)) continue;
-            const frameSpan = frame.openSpan;
-            const msg =
-              frame.kind === 'Repeat'
-                ? `"repeat" without matching "until <cc>"`
-                : `"${frame.kind.toLowerCase()}" without matching "end"`;
-            diag(diagnostics, modulePath, msg, {
-              line: frameSpan.start.line,
-              column: frameSpan.start.column,
-            });
-          }
-          diag(
-            diagnostics,
-            modulePath,
-            `Unterminated func "${name}": expected "end" before "${interruptedByKeyword}"`,
-            { line: interruptedByLine, column: 1 },
-          );
-          continue;
-        }
-        for (const frame of asmControlStack) {
-          if (isRecoverOnlyControlFrame(frame)) continue;
-          const span = frame.openSpan;
-          const msg =
-            frame.kind === 'Repeat'
-              ? `"repeat" without matching "until <cc>"`
-              : `"${frame.kind.toLowerCase()}" without matching "end"`;
-          diag(diagnostics, modulePath, msg, { line: span.start.line, column: span.start.column });
-        }
-        diag(diagnostics, modulePath, `Unterminated func "${name}": missing "end"`, {
-          line: lineNo,
-          column: 1,
-        });
-        break;
-      }
-
+          getRawLine,
+          isReservedTopLevelName,
+          parseParamsFromText,
+        },
+      );
+      if (parsedFunc.node) items.push(parsedFunc.node);
+      i = parsedFunc.nextIndex;
       continue;
     }
